requirements.txt
Django
djangorestframework
django-cors-headers
daphne
psycopg2